[TASK] Use only front-end domain with purge URLs.
When using asynchronous purge queue processing with RealUrl cache tables
the purge URL for a page-id is generated by first looking up in to the
cache tables. If it is found there then fine, else the URL is generated by
appending the page-id with the host domain i.e. backend domain. This might
not be same as front-end domain in some cases. Therefore it makes sense to
only use the front domain with PURGE requests.

Change-Id: If1cf8b5b05c83c6120a964fab8343d8fbad2fbc8
Reviewed-on: http://review.typo3.org/31900
Reviewed-by: Chetan Thapliyal <nospam.ct@gmail.com>
Tested-by: Chetan Thapliyal <nospam.ct@gmail.com>